+++ tests/libYARP_sig/VectorOfTest.cpp
@@ -244,7 +244,7 @@ public:
 
         //compare vector and tmp
         success = true;
-        if (tmp2.size() != (int)vector.size())
+        if (tmp2.size() != vector.size())
         {
             success = false;
         }
